# fe/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VAI Code",
    "description": "Weather Bridge AI production-oriented API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/residents": {
      "get": {
        "tags": [
          "residents"
        ],
        "summary": "List Residents",
        "operationId": "list_residents_api_v1_residents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResidentResponse"
                  },
                  "type": "array",
                  "title": "Response List Residents Api V1 Residents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Create Resident",
        "operationId": "create_resident_api_v1_residents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/{resident_id}/link-account": {
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Link Resident Account",
        "operationId": "link_resident_account_api_v1_residents__resident_id__link_account_post",
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkResidentAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/{resident_id}": {
      "get": {
        "tags": [
          "residents"
        ],
        "summary": "Get Resident",
        "operationId": "get_resident_api_v1_residents__resident_id__get",
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/{resident_id}/contacts": {
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Add Contact",
        "operationId": "add_contact_api_v1_residents__resident_id__contacts_post",
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/{resident_id}/contacts/{contact_id}": {
      "patch": {
        "tags": [
          "residents"
        ],
        "summary": "Update Contact",
        "operationId": "update_contact_api_v1_residents__resident_id__contacts__contact_id__patch",
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          },
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/{resident_id}/locations": {
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Add Location",
        "operationId": "add_location_api_v1_residents__resident_id__locations_post",
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentPointRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentLocationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/residents/{resident_id}/support-needs": {
      "post": {
        "tags": [
          "residents"
        ],
        "summary": "Add Support Need",
        "operationId": "add_support_need_api_v1_residents__resident_id__support_needs_post",
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupportNeedRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactCreateRequest": {
        "properties": {
          "channel": {
            "type": "string",
            "enum": [
              "sms",
              "zalo",
              "email",
              "web_push",
              "webhook"
            ],
            "title": "Channel"
          },
          "value": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Value"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "default": false
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "channel",
          "value"
        ],
        "title": "ContactCreateRequest"
      },
      "ContactResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "masked_value": {
            "type": "string",
            "title": "Masked Value"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "channel",
          "masked_value",
          "is_primary",
          "is_active",
          "verified_at"
        ],
        "title": "ContactResponse"
      },
      "ContactUpdateRequest": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "is_primary": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Primary"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "ContactUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkResidentAccountRequest": {
        "properties": {
          "keycloak_subject": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Keycloak Subject"
          }
        },
        "type": "object",
        "required": [
          "keycloak_subject"
        ],
        "title": "LinkResidentAccountRequest"
      },
...
      "ResidentCreateRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Full Name"
          },
          "village_code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Village Code"
          },
          "birth_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2200.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "source": {
            "type": "string",
            "enum": [
              "self",
              "official",
              "import",
              "demo"
            ],
            "title": "Source",
            "default": "official"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/ContactCreateRequest"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Contacts"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/ResidentPointRequest"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Locations"
          },
          "livelihood_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "farmer",
                  "livestock",
                  "forestry",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Livelihood Type"
          },
          "livelihood_details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Livelihood Details"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "village_code"
        ],
        "title": "ResidentCreateRequest"
      },
      "ResidentDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Profile Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "village_code": {
            "type": "string",
            "title": "Village Code"
          },
          "birth_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "contact_channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contact Channels"
          },
          "livelihood_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Livelihood Types"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/ContactResponse"
            },
            "type": "array",
            "title": "Contacts"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/ResidentLocationResponse"
            },
            "type": "array",
            "title": "Locations"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_profile_id",
          "full_name",
          "village_code",
          "birth_year",
          "verification_status",
          "simulated",
          "contact_channels",
          "livelihood_types",
          "created_at",
          "contacts",
          "locations"
        ],
        "title": "ResidentDetailResponse"
      },
      "ResidentLocationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "location_type": {
            "type": "string",
            "title": "Location Type"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "precision_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precision M"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "location_type",
          "latitude",
          "longitude",
          "label",
          "precision_m",
          "is_active"
        ],
        "title": "ResidentLocationResponse"
      },
      "ResidentPointRequest": {
        "properties": {
          "location_type": {
            "type": "string",
            "enum": [
              "home",
              "farm",
              "livestock",
              "watch_point"
            ],
            "title": "Location Type"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "precision_m": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Precision M"
          }
        },
        "type": "object",
        "required": [
          "location_type",
          "latitude",
          "longitude"
        ],
        "title": "ResidentPointRequest"
      },
      "ResidentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Profile Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "village_code": {
            "type": "string",
            "title": "Village Code"
          },
          "birth_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "contact_channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contact Channels"
          },
          "livelihood_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Livelihood Types"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_profile_id",
          "full_name",
          "village_code",
          "birth_year",
          "verification_status",
          "simulated",
          "contact_channels",
          "livelihood_types",
          "created_at"
        ],
        "title": "ResidentResponse"
      },
...
      "SupportNeedRequest": {
        "properties": {
          "need_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Need Type"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "need_type"
        ],
        "title": "SupportNeedRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}